requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
